window: split glass into a module, keep it across theme changes
Move the platform glass implementation (macOS NSGlassEffectView / Windows Mica / no-op fallback) into a dedicated window_glass module, one file per platform. Mirror the glass-enabled state onto Theme so a full theme swap via apply_config, not just a light/dark mode switch, reapplies the translucent surface adjustments. Sidebar hover now uses an opaque mix so it stays lighter than the active item even over a translucent glass sidebar. Docs synced (en/zh-CN).
