+++ docs/zh-CN/docs/root.md
@@ -77,8 +77,8 @@ window.update(cx, |_, window, cx| {
 })?;
 ```
 
-遵循 Apple Liquid Glass 的原则（玻璃属于导航层，而非内容层），开启后仅导航层表面（`sidebar`、`title_bar`、`tab_bar`）会变为半透明以透出玻璃。玻璃模式下窗口背景本身会变为透明，因此**你的内容区容器必须自己画不透明背景**（如 `bg(cx.theme().background)`）来盖住玻璃。内容与浮层表面（卡片、列表、表格、popover、对话框）保持不透明，以保证内容清晰并避免重影。该行为对应用的所有窗口生效。调用 `window.set_window_glass(false, cx)` 可恢复不透明背景。
+遵循 Apple Liquid Glass 的原则（玻璃属于导航层，而非内容层），开启后仅导航层表面（`sidebar`、`title_bar`、`tab_bar`）会变为半透明以透出玻璃。内容区保持不透明，以保证内容清晰、不透出下方（重影）。玻璃模式下窗口背景本身会变为透明，因此**你的内容区容器必须自己画不透明背景**（如 `bg(cx.theme().background)`）来盖住玻璃。该行为对应用的所有窗口生效。调用 `window.set_window_glass(false, cx)` 可恢复不透明背景。
 
-完整示例参见 [window_glass example](https://github.com/longbridge/gpui-component/tree/main/examples/window_glass)。
+Story 画廊中有可运行的演示——从标题栏的 window glass 操作切换即可。
 
 [Root]: https://docs.rs/gpui-component/latest/gpui_component/root/struct.Root.html
